Update Console->ConsoleOperatorConfig from v1alpha1 to v1
