+++ pkg/apis/console/v1/register.go
